backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
SQLAlchemy
pydantic[email]
pydantic-settings
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pytest
pytest-asyncio
httpx
python-multipart
pandas
alembic
redis
websockets
vaderSentiment==3.3.2